+++ hooks/useZoraMintComment.ts
@@ -1,10 +1,8 @@
 import { useEffect, useState } from "react";
 import useBalance from "./useBalance";
 import { useAccount, usePublicClient, useWriteContract } from "wagmi";
-import {
-  zoraCreator1155ImplABI,
-  zoraCreatorFixedPriceSaleStrategyAddress,
-} from "@zoralabs/protocol-deployments";
+import { zoraCreator1155ImplABI } from "@zoralabs/protocol-deployments";
+import { zoraCreatorFixedPriceSaleStrategyAddress } from "@/lib/protocolSdk/consts";
 import { CHAIN } from "@/lib/consts";
 import {
   Address,
