+++ lib/consts.ts
@@ -14,21 +14,6 @@ import { baseSepolia, base } from "viem/chains";
 export const IS_TESTNET =
   process.env.NEXT_PUBLIC_IS_TESTNET === "true" ? true : false;
 
-export const zoraCreatorFixedPriceSaleStrategyAddress = {
-  1: "0x04E2516A2c207E84a1839755675dfd8eF6302F0a",
-  10: "0x3678862f04290E565cCA2EF163BAeb92Bb76790C",
-  999: "0x04E2516A2c207E84a1839755675dfd8eF6302F0a",
-  8453: "0x04E2516A2c207E84a1839755675dfd8eF6302F0a",
-  42161: "0x1Cd1C1f3b8B779B50Db23155F2Cb244FCcA06B21",
-  81457: "0x3EB144aee170BF62FdA1536e38aF51f08e34A5D0",
-  84532: "0x2994762aA0E4C750c51f333C10d81961faEBE785",
-  421614: "0x1Cd1C1f3b8B779B50Db23155F2Cb244FCcA06B21",
-  7777777: "0x04E2516A2c207E84a1839755675dfd8eF6302F0a",
-  11155111: "0x1Cd1C1f3b8B779B50Db23155F2Cb244FCcA06B21",
-  168587773: "0x3EB144aee170BF62FdA1536e38aF51f08e34A5D0",
-  999999999: "0x6d28164C3CE04A190D5F9f0f8881fc807EAD975A",
-} as const;
-
 // Wagmi
 export const CHAIN = IS_TESTNET ? baseSepolia : base;
 export const CHAIN_ID = CHAIN.id;
